- CHG: Better support for BrAPI v2.1

+++ src/main/java/uk/ac/hutton/ics/brapi/resource/genotyping/map/MarkerPositionSearch.java
@@ -13,7 +13,7 @@ public class MarkerPositionSearch
 {
 	private List<String> linkageGroupNames;
 	private List<String> mapDbIds;
-	private Double       maxPosition;
-	private Double       minPosition;
+	private Long         maxPosition;
+	private Long         minPosition;
 	private List<String> variantDbIds;
 }
